+++ R/dear_seq.R
@@ -76,7 +76,7 @@
 #'
 #'@param nb_cores an integer indicating the number of cores to be used when
 #'\code{parallel_comp} is \code{TRUE}.
-#'Default is \code{parallel::detectCores() - 1}.
+#'Default is \code{ceiling(parallel::detectCores()/2 - 1)}.
 #'
 #'@param preprocessed a logical flag indicating whether the expression data have
 #'already been preprocessed (e.g. log2 transformed). Default is \code{FALSE}, in
@@ -245,7 +245,7 @@ dear_seq <- function(exprmat = NULL, object = NULL,
                      which_test = c("permutation", "asymptotic"),
                      which_weights = c("loclin", "voom", "none"),
                      n_perm = 1000, progressbar = TRUE, parallel_comp = TRUE,
-                     nb_cores = parallel::detectCores() - 1,
+                     nb_cores = ceiling(parallel::detectCores()/2 - 1),
                      preprocessed = FALSE,
                      gene_based_weights = FALSE,
                      bw = "nrd",