+++ R/sp_weights.R
@@ -259,7 +259,6 @@ sp_weights <- function(y, x, phi = NULL, use_phi = TRUE, preprocessed = FALSE,
          "'biweight', 'cosine', 'optcosine'")
   }
   
-  
   if (gene_based) {
     
     w <- function(x) {
@@ -320,7 +319,7 @@ sp_weights <- function(y, x, phi = NULL, use_phi = TRUE, preprocessed = FALSE,
       w <- exp(-stats::approx(x = reverse_trans(smth$x), y = smth$y,
                               xout = reverse_trans(mu_x),
                               rule = 2)$y)
-    }    
+    }
     weights <- matrix(w, nrow(mu_x), ncol(mu_x))
   }
   if(sum(is.na(weights)) < 1){